+++ src/main/java/org/grnet/pidmr/dto/ProviderDto.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
+import org.grnet.pidmr.enums.Validator;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -81,4 +82,13 @@ public class ProviderDto {
     )
     @JsonProperty("relies_on_dois")
     public boolean reliesOnDois;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = Validator.class,
+            description = "The validator of the Provider.",
+            example = "NONE"
+    )
+    @JsonProperty("validator")
+    public String validator;
 }