+++ src/main/java/org/grnet/pidmr/dto/UpdateProviderV2.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
+import org.grnet.pidmr.constraints.StringEnumeration;
+import org.grnet.pidmr.enums.Validator;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -18,4 +20,14 @@ public class UpdateProviderV2 extends UpdateProvider {
     )
     @JsonProperty("resolution_modes")
     public Set<@Valid ResolutionModeRequest> actions = new HashSet<>();
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = Validator.class,
+            description = "The Validator (e.g. NONE, ISBN).",
+            example = "ISBN"
+    )
+    @JsonProperty("validator")
+    @StringEnumeration(enumClass = Validator.class, message = "validator")
+    public String validator;
 }