+++ src/main/java/org/grnet/pidmr/dto/ValidatorResponse.java
@@ -1,0 +1,29 @@
+package org.grnet.pidmr.dto;
+
+public class ValidatorResponse {
+
+    private String name;
+    private String description;
+
+    public ValidatorResponse(String name, String description) {
+        this.name = name;
+        this.description = description;
+    }
+
+    // Getters and setters
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}